requirements.txt
Django>=4.2,<5.0
mysqlclient
Pillow
gunicorn
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
razorpay
django-jazzmin
